% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwtest.R
\name{wwtest}
\alias{wwtest}
\title{Wilcoxon--Wigner homogeneity Test}
\usage{
wwtest(data_matrix)
}
\arguments{
\item{data_matrix}{A numeric symmetric matrix with nrow > 2.}
}
\value{
An object of class "htest" containing the test statistic, leading eigenvalue of the transformed matrix, matrix size, and p-value.
\item{data.name}{The name of the input data matrix.}
\item{statistic}{The test statistic.}
\item{parameter}{The size of the matrix (number of rows/columns).}
\item{p.value}{The p-value associated with the test statistic.}
\item{method}{The name of the test, "Wilcoxon--Wigner Test using Rank Transformation".}
\item{alternative}{The alternative hypothesis: "The entries in the matrix do not come from the same distribution".}
\item{estimate}{The largest eigenvalue of the transformed matrix.}
}
\description{
This function tests whether the independent entries in a symmetric matrix are identically distributed.
It first transforms the entry values to the corresponding normalized ranks,
then computes the first eigenvalue of the transformed matrix. The test statistic is constructed based on
the eigenvalue and compared to the standard normal distribution. The function then returns test result
including the test statistic, alternative hypothesis and p-value.
}
\examples{
mat <- matrix(c(1, 2, 3, 2, 1, 1, 3, 1, 1), 3, 3)
wwtest(mat)
}
